requirements.txt
flask==3.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
